This version is close to final version to be used wiht solps-iter.
Complete implementation of input and output for fluid codes and
set of routines to check particle balances.

however this version is not tested against old version of FACE( FACE1.0)
and is not tested for consistency and particle conservation and heat eq
solver.

Three next coming steps are:
 - testing against old face
 - testing analytical solution for simple problem
 - running with solps-like interface (multiple successive runs with
constant temperature and influx then variables ones)